refactor: extract component state into .svelte.ts files and fix DamageRoll lookup
- Extract state management from Svelte components into separate .svelte.ts files:
  - AttackActionPanel: createAttackPanelState()
  - CastSpellActionPanel: createSpellPanelState()
  - AssessActionDialog: createAssessDialogState()
  - PlayerCharacterHeroicActionsTab: createHeroicActionsTabState()
  - AssessActionPanel: createAssessPanelState()

- Fix DamageRoll._evaluate to properly handle advantage/disadvantage by
  finding the base result with exploded flag first

- Update STYLE_GUIDE.md to change component size guideline from 300 to 500 lines
